feat(ai): add AkiraClaw edge inference API with BLE support on akiraconsole
- Add aiinfer_load/run/unload WASM native API (TFLite Micro backend)
- Gate behind ai.infer capability (AKIRA_CAP_AIINFER = 1ULL << 32,
  requires uint64_t cap_mask expansion done in prior commit)
- Fix WAMR native pointer convention: *~ signatures deliver native
  pointers directly; remove erroneous wasm_runtime_addr_app_to_native
  double-conversion that returned NULL
- Fix model buffer allocation: use akira_malloc_buffer (PSRAM via
  shared_multi_heap) instead of k_malloc (DRAM exhausted with BLE active)
- Enable BLE on akiraconsole: move g_signing_state, g_chars, g_svcs to
  PSRAM via AKIRA_BULK_BSS to recover 5192 bytes of DRAM headroom
- Add CONFIG_SHELL_ARGC_MAX=20 to allow 16-byte fs write chunks
- Add 'app install <path> lfs' shell command for LittleFS-based install
- Auto-detect gzip (.akpkg) in app_manager_install_from_path
- Add __stdout_hook_install weak stub for native_sim + EXTERNAL_LIBC
- Document native function implementation rules in runtime.md

Verified: aiinfer_test (hello_world_float sine model) runs on ESP32-S3,
outputs x=0.000 y=0.026 / x=0.785 y=0.700 / x=1.571 y=0.996 /
x=3.142 y=-0.005, exits PASS. native_sim and akiraconsole both build clean.

+++ src/lib/akpkg.c
@@ -733,7 +733,7 @@ int32_t akpkg_inflate(const uint8_t *gz, size_t gz_len,
         return -EIO;
     }
 
-    return (ssize_t)s.out_pos;
+    return (int32_t)s.out_pos;
 }
 
 size_t akpkg_base64_decode(const char *src, size_t src_len,
